fix: add comment dialog 분리
