+++ src/pages/PostsManagerPage.tsx
@@ -28,7 +28,7 @@ import {
 import { UserModal } from "../components/UserModal"
 
 import { useTags } from "../hooks/useTags"
-import { useSelectedPost } from "../store/selectedPost"
+
 import { useSearchQueryStore } from "../store/searchQuery"
 import { PostDetailDialog } from "../components/PostDetailDialog"
 
@@ -36,6 +36,7 @@ import { useSelectedTag } from "../store/selectedTag"
 import { UserPost } from "../hooks/useUserPosts"
 import { PostRow } from "../components/PostRow"
 import { AddCommentDialog } from "../components/AddCommentDialog"
+import { EditPostDialog } from "../components/dialog/EditPostDialog"
 
 const PostsManager = () => {
   const navigate = useNavigate()
@@ -50,11 +51,10 @@ const PostsManager = () => {
   // const [searchQuery, setSearchQuery] = useState(queryParams.get("search") || "")
   const { searchQuery, setSearchQuery } = useSearchQueryStore()
   // const [selectedPost, setSelectedPost] = useState(null)
-  const { selectedPost, setSelectedPost } = useSelectedPost()
   const [sortBy, setSortBy] = useState(queryParams.get("sortBy") || "")
   const [sortOrder, setSortOrder] = useState(queryParams.get("sortOrder") || "asc")
   const [showAddDialog, setShowAddDialog] = useState(false)
-  const [showEditDialog, setShowEditDialog] = useState(false)
+
   const [newPost, setNewPost] = useState({ title: "", body: "", userId: 1 })
   const [loading, setLoading] = useState(false)
   const { data: tags } = useTags()
@@ -168,22 +168,6 @@ const PostsManager = () => {
       setNewPost({ title: "", body: "", userId: 1 })
     } catch (error) {
       console.error("게시물 추가 오류:", error)
-    }
-  }
-
-  // 게시물 업데이트
-  const updatePost = async () => {
-    try {
-      const response = await fetch(`/api/posts/${selectedPost.id}`, {
-        method: "PUT",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(selectedPost),
-      })
-      const data = await response.json()
-      setPosts(posts.map((post) => (post.id === data.id ? data : post)))
-      setShowEditDialog(false)
-    } catch (error) {
-      console.error("게시물 업데이트 오류:", error)
     }
   }
 
@@ -374,27 +358,7 @@ const PostsManager = () => {
       </Dialog>
 
       {/* 게시물 수정 대화상자 */}
-      <Dialog open={showEditDialog} onOpenChange={setShowEditDialog}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>게시물 수정</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <Input
-              placeholder="제목"
-              value={selectedPost?.title || ""}
-              onChange={(e) => setSelectedPost({ ...selectedPost, title: e.target.value })}
-            />
-            <Textarea
-              rows={15}
-              placeholder="내용"
-              value={selectedPost?.body || ""}
-              onChange={(e) => setSelectedPost({ ...selectedPost, body: e.target.value })}
-            />
-            <Button onClick={updatePost}>게시물 업데이트</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
+      <EditPostDialog />
 
       {/* 댓글 추가 대화상자 */}
       <AddCommentDialog />
